Push opam image to hub before building switches
This allows us to build the compiler images using the published repo
digest as the base image, rather than just using the local hash.
The advantages of this are:

- We can build using BuildKit (which doesn't support local hashes).

- The FROM line in the generated Dockerfile can be used by others.

- If we need to rebuild an image after the build machine has purged the
  base image, we can just download it again from Hub.

The main disadvantage is that we can't build everything locally for
testing; we have to push to hub before the compilers can be built.