add i386 container images to the build matrix
This adds an `i386` 32-bit container image to the build matrix
(which is the "386" platform in Docker parlance).

The i386 images are slightly tricker than others since it is the
first cross target in the image matrix. The i386 base images are
held in a separate base image repository. They also requires the
SHELL and ENTRYPOINT to be set to linux32 in order to have a
i686 personality for the image.

All of this is handled in ocaml-dockerfile and ocaml-version,
so this PR updates the git submodules.

After this PR, the 386 images will exist, but their use in a
multiarch model isnt entirely straightforward, since the
documented "FROM --platform=386 ocurrent/opam" doesnt actually
seem to work.  The client picks the last cached version from
the manifest, which is usually the amd64 version.  However,
explicitly using the i386 image (either via directly using its
image label, or `docker manifest inspect` to pull out the right
sha256 hash, works fine.