build an archive base image using the master distro
This archive is built against the opam-repo branch, and then
copied into a bare Alpine container.

The files in this base image are useful for use with opam2web
and co, since an incremental build against every opam-repo
revision is very fast when using the files from this base image.

In the event of a distfile mutation, we can also refer back to
the historical `archive` tags to find the older distfiles